refactor(test):
- combine test cases for builder in one file
- refactor helper methods for tests and remove test_utils.go as they could be included in builder_test.go
- add missing input and expected files for the tests
- comment test functions
