+++ testFiles/testMessages/jsonLD/standardMessage.json
@@ -1,0 +1,52 @@
+{
+  "Test": {
+    "type": "object",
+    "properties": {
+      "testBool": {
+        "type": "boolean"
+      },
+      "testBytes": {
+        "type": "string"
+      },
+      "testDouble": {
+        "type": "number"
+      },
+      "testFixed32": {
+        "type": "integer"
+      },
+      "testFixed64": {
+        "type": "integer"
+      },
+      "testFloat": {
+        "type": "number"
+      },
+      "testInt32": {
+        "type": "integer"
+      },
+      "testInt64": {
+        "type": "integer"
+      },
+      "testSFixed32": {
+        "type": "integer"
+      },
+      "testSFixed64": {
+        "type": "integer"
+      },
+      "testSInt32": {
+        "type": "integer"
+      },
+      "testSInt64": {
+        "type": "integer"
+      },
+      "testString": {
+        "type": "string"
+      },
+      "testUInt32": {
+        "type": "integer"
+      },
+      "testUInt64": {
+        "type": "integer"
+      }
+    }
+  }
+}
