+++ testFiles/testMessages/protos/multipleMessages.proto
@@ -1,0 +1,18 @@
+syntax = "proto3";
+
+message Empty {
+}
+
+message Test1 {
+  int32 t1 = 1;
+}
+
+message Test2 {
+  int32 t2 = 1;
+  int32 t3 = 2;
+  float t4 = 3;
+}
+
+message Test3 {
+  int32 t1 = 1;
+}
